+++ heksher/clients/async_client.py
@@ -17,6 +17,7 @@ from heksher.setting import MISSING, Setting
 logger = getLogger(__name__)
 
 T = TypeVar('T')
+content_header = {"Content-type": "application/json"}
 
 __all__ = ['AsyncHeksherClient']
 
@@ -74,7 +75,8 @@ class AsyncHeksherClient(V1APIClient, ContextFeaturesMixin, AsyncContextManagerM
             }
             if setting.default_value is not MISSING:
                 declaration_data['default_value'] = setting.default_value
-            response = await self._http_client.put('api/v1/settings/declare', data=orjson.dumps(declaration_data))
+            response = await self._http_client.put('api/v1/settings/declare', content=orjson.dumps(declaration_data),
+                                                   headers=content_header)
             self._handle_declaration_response(setting, response)
 
         while True:
@@ -106,7 +108,8 @@ class AsyncHeksherClient(V1APIClient, ContextFeaturesMixin, AsyncContextManagerM
                 data['cache_time'] = self._last_cache_time.isoformat()
             new_cache_time = datetime.utcnow()
 
-            response = await self._http_client.post('/api/v1/rules/query', data=orjson.dumps(data))
+            response = await self._http_client.post('/api/v1/rules/query', content=orjson.dumps(data),
+                                                    headers=content_header)
             response.raise_for_status()
 
             updated_settings = response.json()['rules']